+++ src/android/com/adobe/phonegap/push/PushInstanceIDListenerService.java
@@ -1,5 +1,6 @@
 package com.adobe.phonegap.push;
 
+import android.content.Intent;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.util.Log;
@@ -14,22 +15,13 @@ import java.io.IOException;
 public class PushInstanceIDListenerService extends InstanceIDListenerService implements PushConstants {
     public static final String LOG_TAG = "PushPlugin_PushInstanceIDListenerService";
 
+    @Override
     public void onTokenRefresh() {
-        // re-register
         SharedPreferences sharedPref = getApplicationContext().getSharedPreferences(COM_ADOBE_PHONEGAP_PUSH, Context.MODE_PRIVATE);
         String senderID = sharedPref.getString(SENDER_ID, "");
         if (!"".equals(senderID)) {
-            try {
-                String token = InstanceID.getInstance(getApplicationContext()).getToken(senderID, GCM);
-
-                // save new token
-                SharedPreferences.Editor editor = sharedPref.edit();
-                editor.putString(REGISTRATION_ID, token);
-                editor.commit();
-            } catch (IOException e) {
-                Log.e(LOG_TAG, e.getLocalizedMessage(), e);
-            }
-
+            Intent intent = new Intent(this, RegistrationIntentService.class);
+                startService(intent);
         }
     }
 }